LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scrollable-overflow-transform-unreachable-region-expected.txt


PASS scrollWidth of display: flow-root; direction: ltr; writing-mode: horizontal-tb; flex-direction: row; flex-wrap: nowrap;
PASS scrollHeight of display: flow-root; direction: ltr; writing-mode: horizontal-tb; flex-direction: row; flex-wrap: nowrap;
PASS scrollWidth of display: flow-root; direction: ltr; writing-mode: vertical-lr; flex-direction: row; flex-wrap: nowrap;
PASS scrollHeight of display: flow-root; direction: ltr; writing-mode: vertical-lr; flex-direction: row; flex-wrap: nowrap;
PASS scrollWidth of display: flow-root; direction: ltr; writing-mode: vertical-rl; flex-direction: row; flex-wrap: nowrap;
PASS scrollHeight of display: flow-root; direction: ltr; writing-mode: vertical-rl; flex-direction: row; flex-wrap: nowrap;
PASS scrollWidth of display: flow-root; direction: rtl; writing-mode: horizontal-tb; flex-direction: row; flex-wrap: nowrap;
PASS scrollHeight of display: flow-root; direction: rtl; writing-mode: horizontal-tb; flex-direction: row; flex-wrap: nowrap;
PASS scrollWidth of display: flow-root; direction: rtl; writing-mode: vertical-lr; flex-direction: row; flex-wrap: nowrap;
PASS scrollHeight of display: flow-root; direction: rtl; writing-mode: vertical-lr; flex-direction: row; flex-wrap: nowrap;
PASS scrollWidth of display: flow-root; direction: rtl; writing-mode: vertical-rl; flex-direction: row; flex-wrap: nowrap;
PASS scrollHeight of display: flow-root; direction: rtl; writing-mode: vertical-rl; flex-direction: row; flex-wrap: nowrap;
PASS scrollWidth of display: flex; direction: ltr; writing-mode: horizontal-tb; flex-direction: row; flex-wrap: nowrap;
PASS scrollHeight of display: flex; direction: ltr; writing-mode: horizontal-tb; flex-direction: row; flex-wrap: nowrap;
PASS scrollWidth of display: flex; direction: ltr; writing-mode: vertical-lr; flex-direction: row; flex-wrap: nowrap;
PASS scrollHeight of display: flex; direction: ltr; writing-mode: vertical-lr; flex-direction: row; flex-wrap: nowrap;
PASS scrollWidth of display: flex; direction: ltr; writing-mode: vertical-rl; flex-direction: row; flex-wrap: nowrap;
PASS scrollHeight of display: flex; direction: ltr; writing-mode: vertical-rl; flex-direction: row; flex-wrap: nowrap;
PASS scrollWidth of display: flex; direction: rtl; writing-mode: horizontal-tb; flex-direction: row; flex-wrap: nowrap;
PASS scrollHeight of display: flex; direction: rtl; writing-mode: horizontal-tb; flex-direction: row; flex-wrap: nowrap;
PASS scrollWidth of display: flex; direction: rtl; writing-mode: vertical-lr; flex-direction: row; flex-wrap: nowrap;
PASS scrollHeight of display: flex; direction: rtl; writing-mode: vertical-lr; flex-direction: row; flex-wrap: nowrap;
PASS scrollWidth of display: flex; direction: rtl; writing-mode: vertical-rl; flex-direction: row; flex-wrap: nowrap;
PASS scrollHeight of display: flex; direction: rtl; writing-mode: vertical-rl; flex-direction: row; flex-wrap: nowrap;
PASS scrollWidth of display: flex; direction: ltr; writing-mode: horizontal-tb; flex-direction: row; flex-wrap: wrap-reverse;
FAIL scrollHeight of display: flex; direction: ltr; writing-mode: horizontal-tb; flex-direction: row; flex-wrap: wrap-reverse; assert_equals: scrollHeight expected 216 but got 204
FAIL scrollWidth of display: flex; direction: ltr; writing-mode: vertical-lr; flex-direction: row; flex-wrap: wrap-reverse; assert_equals: scrollWidth expected 108 but got 102
PASS scrollHeight of display: flex; direction: ltr; writing-mode: vertical-lr; flex-direction: row; flex-wrap: wrap-reverse;
FAIL scrollWidth of display: flex; direction: ltr; writing-mode: vertical-rl; flex-direction: row; flex-wrap: wrap-reverse; assert_equals: scrollWidth expected 102 but got 108
PASS scrollHeight of display: flex; direction: ltr; writing-mode: vertical-rl; flex-direction: row; flex-wrap: wrap-reverse;
PASS scrollWidth of display: flex; direction: rtl; writing-mode: horizontal-tb; flex-direction: row; flex-wrap: wrap-reverse;
FAIL scrollHeight of display: flex; direction: rtl; writing-mode: horizontal-tb; flex-direction: row; flex-wrap: wrap-reverse; assert_equals: scrollHeight expected 216 but got 204
FAIL scrollWidth of display: flex; direction: rtl; writing-mode: vertical-lr; flex-direction: row; flex-wrap: wrap-reverse; assert_equals: scrollWidth expected 108 but got 102
PASS scrollHeight of display: flex; direction: rtl; writing-mode: vertical-lr; flex-direction: row; flex-wrap: wrap-reverse;
FAIL scrollWidth of display: flex; direction: rtl; writing-mode: vertical-rl; flex-direction: row; flex-wrap: wrap-reverse; assert_equals: scrollWidth expected 102 but got 108
PASS scrollHeight of display: flex; direction: rtl; writing-mode: vertical-rl; flex-direction: row; flex-wrap: wrap-reverse;
FAIL scrollWidth of display: flex; direction: ltr; writing-mode: horizontal-tb; flex-direction: row-reverse; flex-wrap: nowrap; assert_equals: scrollWidth expected 108 but got 102
PASS scrollHeight of display: flex; direction: ltr; writing-mode: horizontal-tb; flex-direction: row-reverse; flex-wrap: nowrap;
PASS scrollWidth of display: flex; direction: ltr; writing-mode: vertical-lr; flex-direction: row-reverse; flex-wrap: nowrap;
FAIL scrollHeight of display: flex; direction: ltr; writing-mode: vertical-lr; flex-direction: row-reverse; flex-wrap: nowrap; assert_equals: scrollHeight expected 216 but got 204
PASS scrollWidth of display: flex; direction: ltr; writing-mode: vertical-rl; flex-direction: row-reverse; flex-wrap: nowrap;
FAIL scrollHeight of display: flex; direction: ltr; writing-mode: vertical-rl; flex-direction: row-reverse; flex-wrap: nowrap; assert_equals: scrollHeight expected 216 but got 204
FAIL scrollWidth of display: flex; direction: rtl; writing-mode: horizontal-tb; flex-direction: row-reverse; flex-wrap: nowrap; assert_equals: scrollWidth expected 102 but got 108
PASS scrollHeight of display: flex; direction: rtl; writing-mode: horizontal-tb; flex-direction: row-reverse; flex-wrap: nowrap;
PASS scrollWidth of display: flex; direction: rtl; writing-mode: vertical-lr; flex-direction: row-reverse; flex-wrap: nowrap;
FAIL scrollHeight of display: flex; direction: rtl; writing-mode: vertical-lr; flex-direction: row-reverse; flex-wrap: nowrap; assert_equals: scrollHeight expected 204 but got 216
PASS scrollWidth of display: flex; direction: rtl; writing-mode: vertical-rl; flex-direction: row-reverse; flex-wrap: nowrap;
FAIL scrollHeight of display: flex; direction: rtl; writing-mode: vertical-rl; flex-direction: row-reverse; flex-wrap: nowrap; assert_equals: scrollHeight expected 204 but got 216
FAIL scrollWidth of display: flex; direction: ltr; writing-mode: horizontal-tb; flex-direction: row-reverse; flex-wrap: wrap-reverse; assert_equals: scrollWidth expected 108 but got 102
FAIL scrollHeight of display: flex; direction: ltr; writing-mode: horizontal-tb; flex-direction: row-reverse; flex-wrap: wrap-reverse; assert_equals: scrollHeight expected 216 but got 204
FAIL scrollWidth of display: flex; direction: ltr; writing-mode: vertical-lr; flex-direction: row-reverse; flex-wrap: wrap-reverse; assert_equals: scrollWidth expected 108 but got 102
FAIL scrollHeight of display: flex; direction: ltr; writing-mode: vertical-lr; flex-direction: row-reverse; flex-wrap: wrap-reverse; assert_equals: scrollHeight expected 216 but got 204
FAIL scrollWidth of display: flex; direction: ltr; writing-mode: vertical-rl; flex-direction: row-reverse; flex-wrap: wrap-reverse; assert_equals: scrollWidth expected 102 but got 108
FAIL scrollHeight of display: flex; direction: ltr; writing-mode: vertical-rl; flex-direction: row-reverse; flex-wrap: wrap-reverse; assert_equals: scrollHeight expected 216 but got 204
FAIL scrollWidth of display: flex; direction: rtl; writing-mode: horizontal-tb; flex-direction: row-reverse; flex-wrap: wrap-reverse; assert_equals: scrollWidth expected 102 but got 108
FAIL scrollHeight of display: flex; direction: rtl; writing-mode: horizontal-tb; flex-direction: row-reverse; flex-wrap: wrap-reverse; assert_equals: scrollHeight expected 216 but got 204
FAIL scrollWidth of display: flex; direction: rtl; writing-mode: vertical-lr; flex-direction: row-reverse; flex-wrap: wrap-reverse; assert_equals: scrollWidth expected 108 but got 102
FAIL scrollHeight of display: flex; direction: rtl; writing-mode: vertical-lr; flex-direction: row-reverse; flex-wrap: wrap-reverse; assert_equals: scrollHeight expected 204 but got 216
FAIL scrollWidth of display: flex; direction: rtl; writing-mode: vertical-rl; flex-direction: row-reverse; flex-wrap: wrap-reverse; assert_equals: scrollWidth expected 102 but got 108
FAIL scrollHeight of display: flex; direction: rtl; writing-mode: vertical-rl; flex-direction: row-reverse; flex-wrap: wrap-reverse; assert_equals: scrollHeight expected 204 but got 216
PASS scrollWidth of display: flex; direction: ltr; writing-mode: horizontal-tb; flex-direction: column; flex-wrap: nowrap;
PASS scrollHeight of display: flex; direction: ltr; writing-mode: horizontal-tb; flex-direction: column; flex-wrap: nowrap;
PASS scrollWidth of display: flex; direction: ltr; writing-mode: vertical-lr; flex-direction: column; flex-wrap: nowrap;
PASS scrollHeight of display: flex; direction: ltr; writing-mode: vertical-lr; flex-direction: column; flex-wrap: nowrap;
PASS scrollWidth of display: flex; direction: ltr; writing-mode: vertical-rl; flex-direction: column; flex-wrap: nowrap;
PASS scrollHeight of display: flex; direction: ltr; writing-mode: vertical-rl; flex-direction: column; flex-wrap: nowrap;
PASS scrollWidth of display: flex; direction: rtl; writing-mode: horizontal-tb; flex-direction: column; flex-wrap: nowrap;
PASS scrollHeight of display: flex; direction: rtl; writing-mode: horizontal-tb; flex-direction: column; flex-wrap: nowrap;
PASS scrollWidth of display: flex; direction: rtl; writing-mode: vertical-lr; flex-direction: column; flex-wrap: nowrap;
PASS scrollHeight of display: flex; direction: rtl; writing-mode: vertical-lr; flex-direction: column; flex-wrap: nowrap;
PASS scrollWidth of display: flex; direction: rtl; writing-mode: vertical-rl; flex-direction: column; flex-wrap: nowrap;
PASS scrollHeight of display: flex; direction: rtl; writing-mode: vertical-rl; flex-direction: column; flex-wrap: nowrap;
FAIL scrollWidth of display: flex; direction: ltr; writing-mode: horizontal-tb; flex-direction: column; flex-wrap: wrap-reverse; assert_equals: scrollWidth expected 108 but got 102
PASS scrollHeight of display: flex; direction: ltr; writing-mode: horizontal-tb; flex-direction: column; flex-wrap: wrap-reverse;
PASS scrollWidth of display: flex; direction: ltr; writing-mode: vertical-lr; flex-direction: column; flex-wrap: wrap-reverse;
FAIL scrollHeight of display: flex; direction: ltr; writing-mode: vertical-lr; flex-direction: column; flex-wrap: wrap-reverse; assert_equals: scrollHeight expected 216 but got 204
PASS scrollWidth of display: flex; direction: ltr; writing-mode: vertical-rl; flex-direction: column; flex-wrap: wrap-reverse;
FAIL scrollHeight of display: flex; direction: ltr; writing-mode: vertical-rl; flex-direction: column; flex-wrap: wrap-reverse; assert_equals: scrollHeight expected 216 but got 204
FAIL scrollWidth of display: flex; direction: rtl; writing-mode: horizontal-tb; flex-direction: column; flex-wrap: wrap-reverse; assert_equals: scrollWidth expected 102 but got 108
PASS scrollHeight of display: flex; direction: rtl; writing-mode: horizontal-tb; flex-direction: column; flex-wrap: wrap-reverse;
PASS scrollWidth of display: flex; direction: rtl; writing-mode: vertical-lr; flex-direction: column; flex-wrap: wrap-reverse;
FAIL scrollHeight of display: flex; direction: rtl; writing-mode: vertical-lr; flex-direction: column; flex-wrap: wrap-reverse; assert_equals: scrollHeight expected 204 but got 216
PASS scrollWidth of display: flex; direction: rtl; writing-mode: vertical-rl; flex-direction: column; flex-wrap: wrap-reverse;
FAIL scrollHeight of display: flex; direction: rtl; writing-mode: vertical-rl; flex-direction: column; flex-wrap: wrap-reverse; assert_equals: scrollHeight expected 204 but got 216
PASS scrollWidth of display: flex; direction: ltr; writing-mode: horizontal-tb; flex-direction: column-reverse; flex-wrap: nowrap;
FAIL scrollHeight of display: flex; direction: ltr; writing-mode: horizontal-tb; flex-direction: column-reverse; flex-wrap: nowrap; assert_equals: scrollHeight expected 216 but got 204
FAIL scrollWidth of display: flex; direction: ltr; writing-mode: vertical-lr; flex-direction: column-reverse; flex-wrap: nowrap; assert_equals: scrollWidth expected 108 but got 102
PASS scrollHeight of display: flex; direction: ltr; writing-mode: vertical-lr; flex-direction: column-reverse; flex-wrap: nowrap;
FAIL scrollWidth of display: flex; direction: ltr; writing-mode: vertical-rl; flex-direction: column-reverse; flex-wrap: nowrap; assert_equals: scrollWidth expected 102 but got 108
PASS scrollHeight of display: flex; direction: ltr; writing-mode: vertical-rl; flex-direction: column-reverse; flex-wrap: nowrap;
PASS scrollWidth of display: flex; direction: rtl; writing-mode: horizontal-tb; flex-direction: column-reverse; flex-wrap: nowrap;
FAIL scrollHeight of display: flex; direction: rtl; writing-mode: horizontal-tb; flex-direction: column-reverse; flex-wrap: nowrap; assert_equals: scrollHeight expected 216 but got 204
FAIL scrollWidth of display: flex; direction: rtl; writing-mode: vertical-lr; flex-direction: column-reverse; flex-wrap: nowrap; assert_equals: scrollWidth expected 108 but got 102
PASS scrollHeight of display: flex; direction: rtl; writing-mode: vertical-lr; flex-direction: column-reverse; flex-wrap: nowrap;
FAIL scrollWidth of display: flex; direction: rtl; writing-mode: vertical-rl; flex-direction: column-reverse; flex-wrap: nowrap; assert_equals: scrollWidth expected 102 but got 108
PASS scrollHeight of display: flex; direction: rtl; writing-mode: vertical-rl; flex-direction: column-reverse; flex-wrap: nowrap;
FAIL scrollWidth of display: flex; direction: ltr; writing-mode: horizontal-tb; flex-direction: column-reverse; flex-wrap: wrap-reverse; assert_equals: scrollWidth expected 108 but got 102
FAIL scrollHeight of display: flex; direction: ltr; writing-mode: horizontal-tb; flex-direction: column-reverse; flex-wrap: wrap-reverse; assert_equals: scrollHeight expected 216 but got 204
FAIL scrollWidth of display: flex; direction: ltr; writing-mode: vertical-lr; flex-direction: column-reverse; flex-wrap: wrap-reverse; assert_equals: scrollWidth expected 108 but got 102
FAIL scrollHeight of display: flex; direction: ltr; writing-mode: vertical-lr; flex-direction: column-reverse; flex-wrap: wrap-reverse; assert_equals: scrollHeight expected 216 but got 204
FAIL scrollWidth of display: flex; direction: ltr; writing-mode: vertical-rl; flex-direction: column-reverse; flex-wrap: wrap-reverse; assert_equals: scrollWidth expected 102 but got 108
FAIL scrollHeight of display: flex; direction: ltr; writing-mode: vertical-rl; flex-direction: column-reverse; flex-wrap: wrap-reverse; assert_equals: scrollHeight expected 216 but got 204
FAIL scrollWidth of display: flex; direction: rtl; writing-mode: horizontal-tb; flex-direction: column-reverse; flex-wrap: wrap-reverse; assert_equals: scrollWidth expected 102 but got 108
FAIL scrollHeight of display: flex; direction: rtl; writing-mode: horizontal-tb; flex-direction: column-reverse; flex-wrap: wrap-reverse; assert_equals: scrollHeight expected 216 but got 204
FAIL scrollWidth of display: flex; direction: rtl; writing-mode: vertical-lr; flex-direction: column-reverse; flex-wrap: wrap-reverse; assert_equals: scrollWidth expected 108 but got 102
FAIL scrollHeight of display: flex; direction: rtl; writing-mode: vertical-lr; flex-direction: column-reverse; flex-wrap: wrap-reverse; assert_equals: scrollHeight expected 204 but got 216
FAIL scrollWidth of display: flex; direction: rtl; writing-mode: vertical-rl; flex-direction: column-reverse; flex-wrap: wrap-reverse; assert_equals: scrollWidth expected 102 but got 108
FAIL scrollHeight of display: flex; direction: rtl; writing-mode: vertical-rl; flex-direction: column-reverse; flex-wrap: wrap-reverse; assert_equals: scrollHeight expected 204 but got 216
PASS scrollWidth of display: grid; direction: ltr; writing-mode: horizontal-tb; flex-direction: row; flex-wrap: nowrap;
PASS scrollHeight of display: grid; direction: ltr; writing-mode: horizontal-tb; flex-direction: row; flex-wrap: nowrap;
PASS scrollWidth of display: grid; direction: ltr; writing-mode: vertical-lr; flex-direction: row; flex-wrap: nowrap;
PASS scrollHeight of display: grid; direction: ltr; writing-mode: vertical-lr; flex-direction: row; flex-wrap: nowrap;
PASS scrollWidth of display: grid; direction: ltr; writing-mode: vertical-rl; flex-direction: row; flex-wrap: nowrap;
PASS scrollHeight of display: grid; direction: ltr; writing-mode: vertical-rl; flex-direction: row; flex-wrap: nowrap;
PASS scrollWidth of display: grid; direction: rtl; writing-mode: horizontal-tb; flex-direction: row; flex-wrap: nowrap;
PASS scrollHeight of display: grid; direction: rtl; writing-mode: horizontal-tb; flex-direction: row; flex-wrap: nowrap;
PASS scrollWidth of display: grid; direction: rtl; writing-mode: vertical-lr; flex-direction: row; flex-wrap: nowrap;
PASS scrollHeight of display: grid; direction: rtl; writing-mode: vertical-lr; flex-direction: row; flex-wrap: nowrap;
PASS scrollWidth of display: grid; direction: rtl; writing-mode: vertical-rl; flex-direction: row; flex-wrap: nowrap;
PASS scrollHeight of display: grid; direction: rtl; writing-mode: vertical-rl; flex-direction: row; flex-wrap: nowrap;
